partial merge with release: put back support for old config files (really? Maybe we should just remove it from release); 'fix' the stub config for the sake of hobbled quake readers such as cm3ide (ust earlier build time TARGET to write the file instead of later build time HOST)

+++ scripts/make-bin-dist-min.sh
@@ -28,6 +28,8 @@ DIST="${DIST:-std}" # may be min, core, std, all
 header "building CM3 installation in ${INSTALLROOT}"
 NOCLEAN=${NOCLEAN:-""}
 
+NEWCFG=${NEWCFG:-y}
+
 DS=${DS:-`date -u +'%Y-%m-%d-%H-%M-%S' | tr -d '\\n'`}
 
 # keep short runpaths
@@ -87,19 +89,35 @@ strip_exe "${INSTALLROOT}/bin/mklib${EXE}"
 #-----------------------------------------------------------------------------
 # configure a temporary config file
 echo configuring temporary config file "${INSTALLROOT}/bin/cm3.cfg"
-# delete old config files
-for f in ${ROOT}/m3-sys/cminstall/src/config-no-install/*; do
-  b=`basename ${f}`
-  if [ -f "${INSTALLROOT}/bin/${b}" ] ; then
-    rm "${INSTALLROOT}/bin/${b}" > /dev/null
+if [ "${NEWCFG}" != "y" ]; then
+  # old style installation
+  if [ "${TARGET}" = "NT386" -o "${TARGET}" = "NT386GNU" ]; then
+    CFG1="${ROOT}/m3-sys/cm3/src/config/${TARGET}.main"
+    CFG2="${ROOT}/m3-sys/cminstall/src/config/${TARGET}.main"
+    CFG3="${ROOT}/m3-sys/cminstall/src/config/${TARGET}.common"
+    cp "${CFG3}" "${INSTALLROOT}/bin"
+  else
+    CFG1="${ROOT}/m3-sys/cm3/src/config/${TARGET}"
   fi
- done
-# new config files
-cp "${ROOT}/m3-sys/cminstall/src/config-no-install/"* "${INSTALLROOT}/bin/config"
-(
-  echo "INSTALL_ROOT = path() & \"/..\""
-  echo "include(path() & \"/config/\" & HOST)"
-) > "${INSTALLROOT}/bin/cm3.cfg"
+  echo "adjusting INSTALL_ROOT in old config file"
+  sed -e '
+    /^INSTALL_ROOT[ \t]*=/s;^.*$;INSTALL_ROOT = "'${INSTALLROOT}${SL}'";
+  ' "${CFG1}" > "${INSTALLROOT}/bin/cm3.cfg"
+else
+  # delete old config files
+  for f in ${ROOT}/m3-sys/cminstall/src/config-no-install/*; do
+    b=`basename ${f}`
+    if [ -f "${INSTALLROOT}/bin/${b}" ] ; then
+      rm "${INSTALLROOT}/bin/${b}" > /dev/null
+    fi
+  done
+  # new config files
+  cp "${ROOT}/m3-sys/cminstall/src/config-no-install/"* "${INSTALLROOT}/bin/config"
+  (
+    echo "INSTALL_ROOT = path() & \"/..\""
+    echo "include(path() & \"/config/${TARGET}\")"
+  ) > "${INSTALLROOT}/bin/cm3.cfg"
+fi
 
 #-----------------------------------------------------------------------------
 # clean everything
